+++ backend/app/data_processing/image.py
@@ -1,0 +1,6 @@
+"""
+
+This module contains functions for transforming image input data into quantities
+that we can plug into the parameters of our synthesis modules.
+
+"""
